cc-webapp/backend/requirements.latin1.txt
# FastAPI   
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
#  
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9

# Redis  
redis==5.0.1
hiredis==2.2.3

#   
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Kafka  
kafka-python==2.0.2
aiokafka==0.9.0

#  
celery==5.3.4
flower==2.0.1
apscheduler==3.10.4

#   
structlog==23.2.0
python-json-logger==2.0.7
prometheus-fastapi-instrumentator==6.1.0
sentry-sdk[fastapi]==1.39.1

# HTTP 
httpx==0.27.2
aiohttp==3.9.1
requests==2.32.4

# 
python-dotenv==1.0.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-multipart==0.0.6

#  
pytest==7.4.3
pytest-asyncio
pytest-mock==3.12.0
pytest-cov==4.1.0
coverage==7.3.2
black==23.11.0
flake8==6.1.0

#   
aiosqlite==0.19.0

#  
pytz==2023.3
langdetect
